fix(calls): hotfix 4 ошибок компиляции — (1) VideoTextureRenderer: у SurfaceTexture есть только setter setDefaultBufferSize, геттера defaultBufferSize НЕ существует → размер буфера запоминаем в локальных bufW/bufH и для setter, и для лога; (2) WebRtcEngine fillPlane: литерал 128 не влезает в signed Byte -128..127 (Kotlin вписывает литерал автоматически только если влезает — 16 молча скомпилировалось, 128 нет) → сигнатура value: Int с put(value.toByte()) внутри, 128.toByte()==-128 те же биты 0x80. Поведение не изменено. #CALLS-SYMMETRIC #CALLS-VIDEO-TEXVIEW-2

+++ app/src/main/java/re/pinok/media/WebRtcEngine.kt
@@ -890,9 +890,17 @@ class WebRtcEngine(
         signalingHandler?.postDelayed({ pumpDummyFrame() }, 100L)
     }
 
-    private fun fillPlane(plane: java.nio.ByteBuffer, value: Byte) {
+    /**
+     * Заполнение плоскости YUV константой. Значение задаётся в unsigned-диапазоне
+     * 0..255 и переводится в signed Byte внутри (128.toByte() == -128, те же биты
+     * 0x80 — нейтральная хрома).
+     * ВАЖНО: параметр НЕ Byte — литерал 128 не влезает в signed -128..127, и Kotlin
+     * НЕ конвертирует его автоматически (ошибка «actual Int, expected Byte»);
+     * 16 (Y-чёрный) влезает и компилируется молча — рассинхрон типов.
+     */
+    private fun fillPlane(plane: java.nio.ByteBuffer, value: Int) {
         plane.rewind()
-        while (plane.hasRemaining()) plane.put(value)
+        while (plane.hasRemaining()) plane.put(value.toByte())
     }
 
     private fun stopDummyVideo() {
